# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Радар вузов API
  description: '

    CRM контроля взаимодействия ИТ Школы Ростелекома с вузами.


    Все методы `/api/v1` требуют `Authorization: Bearer <access_token>` из Keycloak.

    Ошибки возвращаются в формате `application/problem+json` с полем `code`

    из каталога в ARCHITECTURE.md.

    '
  version: 0.1.0
paths:
...
  /api/v1/integrations:
    get:
      tags:
      - integrations
      summary: Источники LMS и сайта
      operationId: read_integrations_api_v1_integrations_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/IntegrationSourceOut'
                type: array
                title: Response Read Integrations Api V1 Integrations Get
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - HTTPBearer: []
  /api/v1/integrations/{source_id}/sync:
    post:
      tags:
      - integrations
      summary: Синхронизировать источник вручную
      description: 'Запись в журнале появляется и при отказе источника: там будет код ошибки.'
      operationId: post_sync_api_v1_integrations__source_id__sync_post
      security:
      - HTTPBearer: []
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Source Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncRunOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/integrations/{source_id}:
    patch:
      tags:
      - integrations
      summary: Включить или выключить отправку изменений источнику
      operationId: patch_source_api_v1_integrations__source_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Source Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationSourceOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/integrations/{source_id}/push:
    post:
      tags:
      - integrations
      summary: Отправить изменения записей в систему сейчас
      description: 'То же, что делает воркер раз в минуту: документы `radar-vuzov/interaction@1` по записям
        из очереди уходят одним пакетом. Запуск пишется в журнал с направлением `push`, отказ получателя
        — с кодом `INTEGRATION_UNAVAILABLE` и повтором позже.'
      operationId: post_push_api_v1_integrations__source_id__push_post
      security:
      - HTTPBearer: []
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Source Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncRunOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/integrations/{source_id}/outbox:
    get:
      tags:
      - integrations
      summary: Очередь отправки источнику
      description: Какие записи ждут отправки, ушли или отклонены получателем и почему.
      operationId: read_outbox_api_v1_integrations__source_id__outbox_get
      security:
      - HTTPBearer: []
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Source Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(pending|sent|failed)$
          - type: 'null'
          description: Состояние отправки
          title: Status
        description: Состояние отправки
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OutboxEntryOut'
                title: Response Read Outbox Api V1 Integrations  Source Id  Outbox Get
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/integrations/{source_id}/runs:
    get:
      tags:
      - integrations
      summary: Журнал обмена с источником в обе стороны
      operationId: read_runs_api_v1_integrations__source_id__runs_get
      security:
      - HTTPBearer: []
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Source Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SyncRunOut'
                title: Response Read Runs Api V1 Integrations  Source Id  Runs Get
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/site-applications:
    get:
      tags:
      - integrations
      summary: Заявки с сайта
      description: Без фильтра — все заявки; `match_status=unmatched` — очередь на разбор.
      operationId: read_applications_api_v1_site_applications_get
      security:
      - HTTPBearer: []
      parameters:
      - name: match_status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(matched|unmatched)$
          - type: 'null'
          description: Состояние сопоставления
          title: Match Status
        description: Состояние сопоставления
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SiteApplicationOut'
                title: Response Read Applications Api V1 Site Applications Get
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/site-applications/{application_id}/match:
    post:
      tags:
      - integrations
      summary: Привязать заявку к взаимодействию
      operationId: post_match_api_v1_site_applications__application_id__match_post
      security:
      - HTTPBearer: []
      parameters:
      - name: application_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Application Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationMatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteApplicationOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
...
    ApplicationMatch:
      properties:
        interaction_id:
          type: string
          format: uuid
          title: Interaction Id
          description: Взаимодействие, к которому относится заявка
      type: object
      required:
      - interaction_id
      title: ApplicationMatch
...
    ErrorCode:
      type: string
      enum:
      - VALIDATION_ERROR
      - AUTH_REQUIRED
      - AUTH_FORBIDDEN
      - NOT_FOUND
      - INTERACTION_VERSION_CONFLICT
      - INTERACTION_DUPLICATE
      - WF_TRANSITION_NOT_ALLOWED
      - WF_COMMENT_REQUIRED
      - WF_ATTACHMENT_REQUIRED
      - WF_VERSION_NOT_DRAFT
      - WF_MIGRATION_MAP_INCOMPLETE
      - FILE_TYPE_NOT_ALLOWED
      - FILE_TOO_LARGE
      - IMPORT_MAPPING_INVALID
      - REPORT_TOO_LARGE
      - INTEGRATION_UNAVAILABLE
      - INTERNAL_ERROR
      title: ErrorCode
...
    FieldError:
      properties:
        field:
          type: string
          title: Field
        message:
          type: string
          title: Message
      type: object
      required:
      - field
      - message
      title: FieldError
...
    IntegrationSourceOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        kind:
          type: string
          title: Kind
          description: lms или site
        name:
          type: string
          title: Name
        base_url:
          type: string
          title: Base Url
        is_mock:
          type: boolean
          title: Is Mock
        schedule_cron:
          anyOf:
          - type: string
          - type: 'null'
          title: Schedule Cron
        last_sync_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Sync At
        push_enabled:
          type: boolean
          title: Push Enabled
          description: Система принимает изменения записей CRM
        last_push_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Push At
      type: object
      required:
      - id
      - kind
      - name
      - base_url
      - is_mock
      - schedule_cron
      - last_sync_at
      - push_enabled
      - last_push_at
      title: IntegrationSourceOut
...
    OutboxEntryOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        interaction_id:
          type: string
          format: uuid
          title: Interaction Id
        reason:
          type: string
          title: Reason
          description: 'Что изменилось: created, transition, owner, attachment и др.'
        status:
          type: string
          title: Status
          description: pending, sent или failed
        attempts:
          type: integer
          title: Attempts
        next_attempt_at:
          type: string
          format: date-time
          title: Next Attempt At
        last_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Error
        created_at:
          type: string
          format: date-time
          title: Created At
        sent_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Sent At
      type: object
      required:
      - id
      - interaction_id
      - reason
      - status
      - attempts
      - next_attempt_at
      - last_error
      - created_at
      - sent_at
      title: OutboxEntryOut
...
    Problem:
      properties:
        type:
          type: string
          title: Type
          default: about:blank
        title:
          type: string
          title: Title
        status:
          type: integer
          title: Status
        code:
          $ref: '#/components/schemas/ErrorCode'
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
        trace_id:
          type: string
          title: Trace Id
        errors:
          items:
            $ref: '#/components/schemas/FieldError'
          type: array
          title: Errors
          default: []
      type: object
      required:
      - title
      - status
      - code
      - trace_id
      title: Problem
      description: Тело ответа с ошибкой.
...
    SiteApplicationOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        external_id:
          type: string
          title: External Id
        university_name:
          type: string
          title: University Name
        program_name:
          type: string
          title: Program Name
        contact_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Contact Name
        comment:
          anyOf:
          - type: string
          - type: 'null'
          title: Comment
        received_at:
          type: string
          format: date-time
          title: Received At
        match_status:
          type: string
          title: Match Status
          description: matched или unmatched
        interaction_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Interaction Id
      type: object
      required:
      - id
      - external_id
      - university_name
      - program_name
      - contact_name
      - comment
      - received_at
      - match_status
      - interaction_id
      title: SiteApplicationOut
    SourceUpdate:
      properties:
        push_enabled:
          type: boolean
          title: Push Enabled
          description: Отправлять ли системе изменения записей
      type: object
      required:
      - push_enabled
      title: SourceUpdate
...
    SyncRunOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        source_id:
          type: string
          format: uuid
          title: Source Id
        direction:
          type: string
          title: Direction
          description: pull — забрали данные, push — отправили изменения
        started_at:
          type: string
          format: date-time
          title: Started At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
        status:
          type: string
          title: Status
          description: running, done или failed
        stats:
          additionalProperties: true
          type: object
          title: Stats
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
      type: object
      required:
      - id
      - source_id
      - direction
      - started_at
      - finished_at
      - status
      - stats
      - error_code
      title: SyncRunOut
...
  securitySchemes:
    HTTPBearer:
      type: http
      description: Access token из Keycloak
      scheme: bearer